sra_explorer/pyproject.toml
[tool.poetry]
name = "sra_explorer"
version = "0.1.0"
description = ""
authors = ["Kamil Burkiewicz <kamil.burkiewicz.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
biopython = "^1.85"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
